More refactoring and move search logic into search.py
