+++ .pre-commit-config.yaml
@@ -10,7 +10,7 @@ repos:
       - id: check-added-large-files
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.10.0
+    rev: v0.11.2
     hooks:
       - id: ruff
         name: lint with ruff
