+++ pyproject.toml
@@ -24,4 +24,8 @@ addopts = ["--ignore=data/"]
 line-length = 100
 indent-width = 4
 target-version = "py312"
-exclude = ["migrations/*"]
+exclude = ["migrations/*", "data/*"]
+
+[tool.flake8]
+max-line-length = 100
+exclude = "migrations/*, data/*"
