+++ src/tangerine/models/__init__.py
@@ -1,0 +1,5 @@
+# import all models
+from .agent import Agent
+from .interactions import Interaction, QuestionEmbedding, RelevanceScore, UserFeedback
+
+__all__ = ["Agent", "RelevanceScore", "QuestionEmbedding", "UserFeedback", "Interaction"]
